move the dumps to another module
